+++ src/hauschmxdiffusion2D/diffusion2d.py
@@ -7,7 +7,7 @@ Example acquired from https://scipython.com/book/chapter-7-matplotlib/examples/t
 import numpy as np
 import matplotlib.pyplot as plt
 
-from output import output_plots
+from .output import output_plots
 
 # plate size, mm
 w = h = 10.
